+++ src/Feed/Adapter/Memory.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Utopia\Feed\Adapter;
 
 use Utopia\Feed\Adapter;
-use Utopia\Feed\Event;
+use Utopia\CloudEvents\CloudEvent;
 use Utopia\Feed\Id;
 
 /**
@@ -21,7 +21,7 @@ use Utopia\Feed\Id;
  */
 class Memory extends Adapter
 {
-    /** @var list<Event> */
+    /** @var list<CloudEvent> */
     private array $events = [];
 
     /**
@@ -38,7 +38,7 @@ class Memory extends Adapter
         parent::__construct($name);
     }
 
-    public function append(Event $event): string
+    public function append(CloudEvent $event): string
     {
         $now = (int) \floor(\microtime(true) * 1000);